# docs/spec/recorder-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Локальный API регистратора JA3Proxy",
    "version": "0.5.0",
    "description": "Наблюдения в ограниченной памяти и явные device mappings. Loopback binding работает без обязательных TLS и Bearer-аутентификации; локальный token admin API разрешён без токена. Для non-loopback binding обязательны HTTPS и Bearer-токен, управление токенами требует роль admin. Реестр секретов хранится в общей SQLite в виде SHA-256."
  },
  "security": [
    {"BearerAuth": []},
    {}
  ],
  "servers": [
    {
      "url": "http://127.0.0.1:9090"
    }
  ],
  "paths": {
    "/api/v1/admin/tokens": {
      "get": {
        "operationId": "listPanelTokens",
        "description": "Возвращает метаданные token registry без bearer-секретов. Для non-loopback требуется роль admin; на loopback авторизация необязательна.",
        "responses": {
          "200": {"description": "Список токенов", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenList"}}}},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "401": {"description": "Bearer-токен отсутствует или недействителен"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      },
      "post": {
        "operationId": "issuePanelToken",
        "description": "Создаёт bearer-токен со случайным секретом. Секрет возвращается только в этом ответе; в SQLite хранится SHA-256.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenIssue"}}}},
        "responses": {
          "201": {"description": "Токен создан; сохраните секрет из ответа", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IssuedToken"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/api/v1/admin/tokens/{id}": {
      "put": {
        "operationId": "updatePanelTokenRole",
        "description": "Меняет роль/scopes/срок действия существующего токена без изменения секрета. Последний активный admin не может быть понижен или лишён роли.",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenUpdate"}}}},
        "responses": {
          "200": {"description": "Роль и разрешения обновлены", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenMetadata"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      },
      "delete": {
        "operationId": "revokePanelToken",
        "description": "Немедленно отзывает токен. Нельзя отозвать последний активный admin-токен.",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "200": {"description": "Токен отозван", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenMetadata"}}}},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/api/v1/admin/tokens/{id}/rotate": {
      "post": {
        "operationId": "rotatePanelToken",
        "description": "Немедленно заменяет секрет активного токена, сохраняя ID, роль и scopes. Новый секрет возвращается только один раз.",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "200": {"description": "Секрет токена заменён", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IssuedToken"}}}},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/api/config": {
      "put": {
        "operationId": "updateRuntimeConfig",
        "description": "Атомарно обновляет runtime-настройки. expected_version обязателен; устаревшая версия получает 409 без применения изменения.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RuntimeConfigUpdate"}}}
        },
        "responses": {
          "200": {"description": "Настройки применены к новым соединениям", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RuntimeConfigResponse"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "409": {"$ref": "#/components/responses/Conflict"}
        }
      }
    },
    "/api/v1/export/config": {
      "get": {
        "operationId": "exportConfigBackup",
        "description": "Экспортирует control-plane конфигурацию в JSON или ZIP. Телеметрия не включается; пароли из upstream URL удаляются перед выдачей.",
        "parameters": [
          {"name": "format", "in": "query", "schema": {"type": "string", "enum": ["json", "zip"], "default": "json"}}
        ],
        "responses": {
          "200": {
            "description": "Файл резервной копии конфигурации",
            "content": {
              "application/json": {"schema": {"$ref": "#/components/schemas/ConfigBackup"}},
              "application/zip": {"schema": {"type": "string", "format": "binary"}}
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/api/v1/import/config/validate": {
      "post": {
        "operationId": "validateConfigBackup",
        "description": "Проверяет JSON или ZIP backup без изменения runtime и файлов. Проверяются схемы registry, TLS profiles, routes и upstream TLS.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {"schema": {"$ref": "#/components/schemas/ConfigBackup"}},
            "application/zip": {"schema": {"type": "string", "format": "binary"}}
          }
        },
        "responses": {
          "200": {"description": "Backup совместим", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigValidation"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "422": {"description": "Backup прочитан, но не прошёл проверку", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigValidation"}}}}
        }
      }
    },
    "/api/v1/export/telemetry": {
      "get": {
        "operationId": "exportTelemetrySQLite",
        "description": "Создаёт согласованный standalone snapshot durable SQLite telemetry. Endpoint требует scopes export и raw; memory-only recorder не поддерживается.",
        "responses": {
          "200": {
            "description": "SQLite-файл telemetry backup",
            "content": {"application/vnd.sqlite3": {"schema": {"type": "string", "format": "binary"}}}
          },
          "403": {"$ref": "#/components/responses/Forbidden"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/api/v1/import/telemetry": {
      "post": {
        "operationId": "importTelemetry",
        "description": "Валидирует SQLite snapshot и идемпотентно объединяет observations с активной durable базой. Активный файл SQLite не заменяется.",
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.sqlite3": {"schema": {"type": "string", "format": "binary"}},
            "application/octet-stream": {"schema": {"type": "string", "format": "binary"}}
          }
        },
        "responses": {
          "200": {
            "description": "Snapshot импортирован",
            "content": {"application/json": {"schema": {"type": "object", "required": ["imported"], "properties": {"imported": {"type": "integer", "minimum": 0}}}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "415": {"description": "Неподдерживаемый Content-Type", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "422": {"description": "SQLite snapshot не прошёл проверку", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/api/v1/import/config": {
      "post": {
        "operationId": "importConfigBackup",
        "description": "Применяет предварительно проверенный JSON или ZIP backup. Все четыре expected_*_version обязательны; runtime settings намеренно не восстанавливаются, поскольку credentials отсутствуют в backup.",
        "parameters": [
          {"name": "expected_devices_version", "in": "query", "required": true, "schema": {"type": "integer", "minimum": 0}},
          {"name": "expected_tls_profiles_version", "in": "query", "required": true, "schema": {"type": "integer", "minimum": 0}},
          {"name": "expected_routes_version", "in": "query", "required": true, "schema": {"type": "integer", "minimum": 0}},
          {"name": "expected_upstream_tls_version", "in": "query", "required": true, "schema": {"type": "integer", "minimum": 0}}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {"schema": {"$ref": "#/components/schemas/ConfigBackup"}},
            "application/zip": {"schema": {"type": "string", "format": "binary"}}
          }
        },
        "responses": {
          "200": {"description": "Компоненты backup применены", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigImportResult"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "422": {"description": "Backup не прошёл preflight", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigValidation"}}}},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/api/v1/audit": {
      "get": {
        "operationId": "listAuditEvents",
        "description": "Постраничный локальный просмотр append-only событий панели. Тела запросов не сохраняются; потенциальные секреты в значениях заменяются на [REDACTED].",
        "parameters": [
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 50}},
          {"name": "cursor", "in": "query", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Страница аудита", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuditPage"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Bearer registry загружается из --web-panel-token-file при первом старте и сохраняется в общей SQLite (только SHA-256 секретов). Поддерживаются роли viewer, operator, investigator, admin, scopes и необязательный expires_at в RFC3339; ротация и отзыв применяются без перезапуска."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Некорректный запрос",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "Conflict": {
        "description": "Устаревшая config_version",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "Forbidden": {
        "description": "Требуется локальный доступ",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "NotFound": {
        "description": "Устройство не найдено",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "ServiceUnavailable": {
        "description": "Durable recorder storage недоступно",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "RuntimeConfigUpdate": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expected_version"],
        "properties": {
          "expected_version": {"type": "integer", "minimum": 0},
          "tlsFingerprint": {"type": "string"},
          "upstream": {"type": "string"},
          "proxyPort": {"type": "integer", "minimum": 1, "maximum": 65535},
          "proxyProtocol": {"enum": ["mixed", "http", "socks5"]},
          "proxyAuthEnabled": {"type": "boolean"},
          "proxyUsername": {"type": "string"},
          "proxyPassword": {"type": "string", "writeOnly": true}
        }
      },
      "TokenIssue": {
        "type": "object",
        "additionalProperties": false,
        "required": ["role"],
        "properties": {
          "id": {"type": "string", "maxLength": 64, "pattern": "^[A-Za-z0-9._:-]+$"},
          "role": {"type": "string", "enum": ["viewer", "operator", "investigator", "admin"]},
          "scopes": {"type": "array", "description": "Необязательно; по умолчанию scopes выбранной роли. Роль admin требует полный набор admin scopes.", "items": {"type": "string", "enum": ["read", "write", "raw", "export"]}},
          "expires_at": {"type": "string", "description": "Необязательная дата RFC3339; пустая строка означает отсутствие срока."}
        }
      },
      "TokenUpdate": {
        "type": "object",
        "additionalProperties": false,
        "required": ["role"],
        "properties": {
          "role": {"type": "string", "enum": ["viewer", "operator", "investigator", "admin"]},
          "scopes": {"type": "array", "description": "Необязательно; по умолчанию scopes выбранной роли. Роль admin требует полный набор admin scopes.", "items": {"type": "string", "enum": ["read", "write", "raw", "export"]}},
          "expires_at": {"type": "string", "description": "Новая дата RFC3339 или пустая строка для бессрочного токена."}
        }
      },
      "TokenMetadata": {
        "type": "object",
        "required": ["id", "role", "scopes", "revoked", "created_at", "updated_at"],
        "properties": {
          "id": {"type": "string"}, "role": {"type": "string"},
          "scopes": {"type": "array", "items": {"type": "string"}},
          "expires_at": {"type": "string", "format": "date-time"},
          "revoked": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "TokenList": {
        "type": "object",
        "required": ["tokens"],
        "properties": {"tokens": {"type": "array", "items": {"$ref": "#/components/schemas/TokenMetadata"}}}
      },
      "IssuedToken": {
        "type": "object",
        "required": ["token", "entry"],
        "properties": {
          "token": {"type": "string", "description": "Bearer-секрет, возвращаемый только один раз при создании или ротации; повторно прочитать его нельзя."},
          "entry": {"$ref": "#/components/schemas/TokenMetadata"}
        }
      },
      "RuntimeConfigResponse": {
        "type": "object",
        "properties": {
          "runtime": {
            "type": "object",
            "properties": {
              "configVersion": {"type": "integer", "minimum": 0},
              "proxyListen": {"type": "string"},
              "proxyPort": {"type": "integer"},
              "proxyProtocol": {"type": "string"},
              "tlsClient": {"type": "string"},
              "tlsVersion": {"type": "string"},
              "upstream": {"type": "string"},
              "upstreamEnabled": {"type": "boolean"},
              "proxyAuthEnabled": {"type": "boolean"},
              "proxyUsername": {"type": "string"},
              "mitmCaCertificateStatus": {"type": "string", "enum": ["UNAVAILABLE", "VALID", "EXPIRING", "EXPIRED", "NOT_YET_VALID"]},
              "mitmCaCertificateNotBefore": {"type": "string", "format": "date-time"},
              "mitmCaCertificateNotAfter": {"type": "string", "format": "date-time"},
              "mitmCaCertificateDaysRemaining": {"type": "integer"},
              "panelCertificateStatus": {"type": "string", "enum": ["UNAVAILABLE", "VALID", "EXPIRING", "EXPIRED", "NOT_YET_VALID"]},
              "panelCertificateNotBefore": {"type": "string", "format": "date-time"},
              "panelCertificateNotAfter": {"type": "string", "format": "date-time"},
              "panelCertificateDaysRemaining": {"type": "integer"}
            },
            "required": ["configVersion"]
          }
        },
        "required": ["runtime"]
      },
      "ConfigBackup": {
        "type": "object",
        "required": ["schema_version", "exported_at", "runtime", "devices", "tls_profiles", "routes", "upstream_tls"],
        "properties": {
          "schema_version": {"const": "ja3proxy-config-backup/1"},
          "exported_at": {"type": "string", "format": "date-time"},
          "runtime": {"type": "object"},
          "devices": {"$ref": "#/components/schemas/DeviceRegistry"},
          "tls_profiles": {"$ref": "#/components/schemas/TLSProfileLibrary"},
          "routes": {"$ref": "#/components/schemas/RouteConfig"},
          "route_config_version": {"type": "integer", "minimum": 0},
          "upstream_tls": {"$ref": "#/components/schemas/UpstreamTLSConfig"},
          "upstream_tls_config_version": {"type": "integer", "minimum": 0}
        }
      },
      "AuditPage": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {"type": "array", "items": {"$ref": "#/components/schemas/AuditEvent"}},
          "next_cursor": {"type": "string"}
        }
      },
      "ConfigValidation": {
        "type": "object",
        "required": ["valid", "schema_version", "components"],
        "properties": {
          "valid": {"type": "boolean"},
          "schema_version": {"type": "string"},
          "components": {"type": "array", "items": {"type": "string"}},
          "errors": {"type": "array", "items": {"type": "string"}}
        }
      },
      "ConfigImportResult": {
        "type": "object",
        "required": ["applied"],
        "properties": {
          "applied": {"type": "array", "items": {"type": "string"}},
          "skipped": {"type": "array", "items": {"type": "string"}},
          "warnings": {"type": "array", "items": {"type": "string"}},
          "config_versions": {"type": "object", "additionalProperties": {"type": "integer", "minimum": 0}}
        }
      },
      "AuditEvent": {
        "type": "object",
        "required": ["schema_version", "id", "timestamp", "actor", "action", "object", "result"],
        "properties": {
          "schema_version": {"const": "audit-event/1"},
          "id": {"type": "string"},
          "timestamp": {"type": "string", "format": "date-time"},
          "actor": {"type": "string"},
          "action": {"type": "string"},
          "object": {"type": "string"},
          "old_value": {"type": "object"},
          "new_value": {"type": "object"},
          "source_ip": {"type": "string"},
          "result": {"type": "string", "enum": ["requested", "success", "rejected"]},
          "prev_hash": {"type": "string", "description": "SHA-256 hash предыдущего события"},
          "hash": {"type": "string", "description": "SHA-256 hash канонического события"}
        }
      },
      "RouteConfig": {
        "type": "object",
        "required": ["rules"],
        "properties": {
          "default": {"type": "object"},
          "rules": {"type": "array", "items": {"type": "object"}}
        }
      },
      "UpstreamTLSConfig": {
        "type": "object",
        "required": ["default", "routes"],
        "properties": {
          "default": {"type": "object"},
          "routes": {"type": "array", "items": {"type": "object"}}
        }
      },
...
      "DeviceRegistry": {
        "type": "object",
        "required": ["schema_version", "config_version", "devices"],
        "properties": {
          "schema_version": {"const": "device-registry/1"},
          "config_version": {"type": "integer", "minimum": 0},
          "devices": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/Device"}
          },
          "assignments": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/ApplicationAssignment"}
          },
          "applications": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/Application"}
          }
        }
      },
      "Device": {
        "type": "object",
        "required": ["id", "name", "enabled"],
        "properties": {
          "id": {"type": "string"},
          "name": {"type": "string"},
          "family_id": {"type": "string", "maxLength": 64, "pattern": "^[A-Za-z0-9._:-]*$", "description": "Необязательная явная связь с другими TLS-профилями того же семейства."},
          "description": {"type": "string"},
          "proxy_username": {"type": "string"},
          "source_ips": {"type": "array", "items": {"type": "string", "format": "ipvanyaddress"}},
          "tags": {"type": "array", "items": {"type": "string"}},
          "platform": {"type": "string"},
          "app_name": {"type": "string"},
          "app_version": {"type": "string"},
          "os_name": {"type": "string"},
          "os_version": {"type": "string"},
          "hardware": {"type": "string"},
          "created_at": {"type": "string"},
          "last_seen_at": {"type": "string"},
          "enabled": {"type": "boolean"}
        }
      },
...
      "Application": {
        "type": "object",
        "required": ["id", "name", "enabled"],
        "properties": {
          "id": {"type": "string"},
          "name": {"type": "string"},
          "publisher": {"type": "string"},
          "platform": {"type": "string"},
          "package": {"type": "string"},
          "description": {"type": "string"},
          "created_at": {"type": "string", "format": "date-time"},
          "enabled": {"type": "boolean"}
        }
      },
...
      "TLSProfileLibrary": {
        "type": "object",
        "properties": {
          "schema_version": {"const": "tls-profile-template/1"},
          "config_version": {"type": "integer"},
          "active_id": {"type": "string"},
          "active_ids": {"type": "array", "items": {"type": "string"}},
          "templates": {"type": "array", "items": {"$ref": "#/components/schemas/TLSProfile"}}
        },
        "required": ["schema_version", "config_version", "templates"]
      },
...
    }
  }
}